feat: export the component an the function
