+++ export/download.js
@@ -1,0 +1,6 @@
+import html2pdf from 'html2pdf.js';
+
+export default function (fileName = 'Pdf') {
+  const elementToPrint = document.querySelector('#element-to-print');
+  html2pdf().from(elementToPrint).save(fileName);
+}
